requirements.txt
asgiref==3.6.0
autopep8==2.0.2
beautifulsoup4==4.12.2
bleach==6.0.0
cached-property==1.5.2
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
cobble==0.1.3
colorama==0.4.6
cryptography==40.0.2
defusedxml==0.7.1
Django==4.1.7
django-js-asset==2.0.0
django-markdownfield==0.10.0
django-mdeditor==0.1.20
dnspython==2.3.0
et-xmlfile==1.1.0
exchangelib==5.0.3
idna==3.4
isodate==0.6.1
lxml==4.9.2
mammoth==1.5.1
Markdown==3.3.4
markdownify==0.11.6
mypy-extensions==1.0.0
mysqlclient==2.1.1
oauthlib==3.2.2
openpyxl==3.1.2
packaging==23.1
pathspec==0.11.1
Pillow==9.5.0
platformdirs==3.2.0
pycodestyle==2.10.0
pycparser==2.21
Pygments==2.15.1
pyspnego==0.9.0
requests==2.30.0
requests-ntlm==1.2.0
requests-oauthlib==1.3.1
shortuuid==1.0.11
six==1.16.0
soupsieve==2.4.1
sqlparse==0.4.3
tomli==2.0.1
tzdata==2023.2
tzlocal==5.0.1
urllib3==2.0.2
webencodings==0.5.1
